+++ test/jsii.test.ts
@@ -1,0 +1,41 @@
+import { JsiiProject } from '../src';
+import { mkdtemp, synthSnapshot } from './util';
+
+describe('author', () => {
+  test('authorEmail and authorAddress can be the same value', () => {
+    const project = new JsiiProject({
+      authorAddress: 'hello@hello.com',
+      authorEmail: 'hello@hello.com',
+      repositoryUrl: 'https://github.com/foo/bar.git',
+      author: 'My Name',
+      name: 'project',
+      outdir: mkdtemp(),
+    });
+
+    const pkgjson = synthSnapshot(project)['package.json'];
+    expect(pkgjson.author).toStrictEqual({
+      email: 'hello@hello.com',
+      name: 'My Name',
+      organization: false,
+    });
+  });
+
+  test('authorUrl and authorAddress can be the same value', () => {
+    const project = new JsiiProject({
+      authorAddress: 'https://foo.bar',
+      authorUrl: 'https://foo.bar',
+      repositoryUrl: 'https://github.com/foo/bar.git',
+      author: 'My Name',
+      name: 'project',
+      outdir: mkdtemp(),
+    });
+
+    const pkgjson = synthSnapshot(project)['package.json'];
+    expect(pkgjson.author).toStrictEqual({
+      name: 'My Name',
+      organization: false,
+      url: 'https://foo.bar',
+    });
+  });
+});
+