+++ test/subproject.test.ts
@@ -1,8 +1,8 @@
 import * as path from 'path';
-import { chdir, cwd } from 'process';
 import * as fs from 'fs-extra';
 import { Project, TextFile, LogLevel, ProjectOptions } from '../src';
-import { mkdtemp, TestProject } from './util';
+import { PROJEN_MARKER } from '../src/common';
+import { TestProject } from './util';
 
 test('composing projects declaratively', () => {
   const comp = new TestProject();
@@ -47,17 +47,6 @@ test('multiple levels', () => {
   expect(child2.outdir).toEqual(path.join(root.outdir, 'child1', 'child2'));
 });
 
-test('outdir="." can only be used if projenrc.js is present in the same directory (to protect against override)', () => {
-  const workdir = mkdtemp();
-  const restore = cwd();
-  chdir(workdir);
-  try {
-    expect(() => new TestSubproject({ name: 'bam', outdir: '.' })).toThrow(/cannot use outdir="\." because projenrc\.js does not exist in the current directory/);
-  } finally {
-    chdir(restore);
-  }
-});
-
 test('subprojects cannot introduce files that override each other', () => {
   const root = new TestProject();
   const child = new TestSubproject({ name: 'sub-project', parent: root, outdir: 'sub-project' });
@@ -70,6 +59,40 @@ test('"outdir" for subprojects must be relative', () => {
   const root = new TestProject();
   expect(() => new TestSubproject({ name: 'foobar', parent: root, outdir: '/foo/bar' })).toThrow(/"outdir" must be a relative path/);
 });
+
+test('subproject generated files do not get cleaned up by parent project', () => {
+  const root = new TestProject();
+  const child = new PreSynthProject({ parent: root, outdir: 'sub-project' });
+
+  // no files have been generated yet
+  expect(fs.existsSync(child.file.absolutePath)).toEqual(false);
+
+  // generate all project files at least once
+  root.synth();
+  expect(child.fileExistedDuringPresynth).toEqual(false);
+  expect(fs.existsSync(child.file.absolutePath)).toEqual(true);
+
+  // resynthesize projects with all generated files already existing
+  root.synth();
+  expect(child.fileExistedDuringPresynth).toEqual(true);
+  expect(fs.existsSync(child.file.absolutePath)).toEqual(true);
+});
+
+// a project that depends on generated files during preSynthesize()
+class PreSynthProject extends Project {
+  public file: TextFile;
+  public fileExistedDuringPresynth: boolean;
+  constructor(options: Omit<ProjectOptions, 'name'> = {}) {
+    super({ name: 'presynth-project', clobber: false, ...options });
+
+    this.file = new TextFile(this, 'presynth.txt', { lines: [PROJEN_MARKER] });
+    this.fileExistedDuringPresynth = false;
+  }
+
+  preSynthesize() {
+    this.fileExistedDuringPresynth = fs.existsSync(this.file.absolutePath);
+  }
+}
 
 class TestSubproject extends Project {
   constructor(options: ProjectOptions) {